LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGLength-rem-expected.txt


FAIL rem unit in SVGLength assert_equals: expected 100 but got 225
PASS Convert back to rem from new user unit value
